Use unsigned menu help current value

+++ src/MenuUtil.cpp
@@ -525,7 +525,7 @@ void CMenuPcs::DrawHelpMessageUS(int msgNo, CFont* font, int, int, _GXColor colo
 
 				if (ChkEquipActive(static_cast<int>(*reinterpret_cast<short*>(menuState + 0x28)) +
 				                   static_cast<int>(*reinterpret_cast<short*>(menuState + 0x34))) != 0) {
-					u16 currentValue = 0;
+					unsigned int currentValue = 0;
 					if (currentItem != -1) {
 						currentValue = *reinterpret_cast<u16*>(Game.unkCFlatData0[2] + currentItem * 0x48 + 6);
 					}
